Actualizando Citas

+++ src/app/views/citas/citas.component.ts
@@ -4,6 +4,7 @@ import {CitaService} from '../../services/cita.service';
 import { NgForm } from '@angular/forms';
 import {FormsModule} from '@angular/forms';
 import {BrowserModule} from '@angular/platform-browser';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cita',
@@ -12,6 +13,7 @@ import {BrowserModule} from '@angular/platform-browser';
 })
 export class CitasComponent implements OnInit {
   citaList: Cita[];
+  ObjetoCita:Cita[];
   Num:number;
   Fecha:string;
   Hora:string;
@@ -33,11 +35,20 @@ export class CitasComponent implements OnInit {
           x["$key"] = element.key;
           this.citaList.push(x as Cita);
         });
+        this.Num = this.citaList.length + 1;
       });
   }
   onSubmit(){
+    this.Estado = "Sin verificar";
     this.citaservice.insertdatos(this.Num,this.Fecha,this.Hora,this.Motivo,this.Estado,this.Mascota,this.Propietario);
-  this.resetForm();
+    Swal.fire({
+      position: 'center',
+      icon: 'success',
+      title: 'Cita Ingresada', 
+      showConfirmButton: false,
+      timer: 1500
+    })
+    this.resetForm();
   }
   resetForm() {
     this.Num =0;
@@ -49,8 +60,28 @@ export class CitasComponent implements OnInit {
     this.Propietario = '';
     
     }
+
+
     onDelete($key: string) {
-      this.citaservice.deleteCita($key);    
+      Swal.fire({
+        title: '¿Estás seguro de eliminar?',
+        text: "Una vez eliminado no podrás recuperarlo",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Eliminar',
+        cancelButtonText: 'Cancelar'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.citaservice.deleteCita($key);
+          Swal.fire(
+            '¡Eliminado!',
+            'El registro ha sido eliminado.',
+            'success'
+          )
+        }
+      })  
   }
 
 }
